providers/proxy: add support for traefik.io API and CRD
